Moved NumericOperator enum to FilterNumeric
