--- Sources/AlgoliaSearch-Client/FilterBuilder/Auxiliary/NumericOperator.swift
@@ -1,18 +1,0 @@
-//
-//  NumericOperator.swift
-//  AlgoliaSearch
-//
-//  Created by Guy Daher on 14/12/2018.
-//  Copyright © 2018 Algolia. All rights reserved.
-//
-
-import Foundation
-
-public enum NumericOperator: String {
-    case lessThan = "<"
-    case lessThanOrEqual = "<="
-    case equals = "="
-    case notEquals = "!="
-    case greaterThanOrEqual = ">="
-    case greaterThan = ">"
-}
